refactor: split advanced-start tables and unit-role SoT keys (#4626)
Topic-split mixed advanced-start tables and key civilian roles with kUnitType constants.
