+++ packages/colonizethis_data/lib/colonizethis_data.dart
@@ -2,7 +2,10 @@
 /// Topology and tile map formats: SPEC/program/map-data.md.
 library colonizethis_data;
 
-export 'src/advanced_start_tables.dart';
+export 'src/advanced_start_tech_tables.dart';
+export 'src/advanced_start_force_tables.dart';
+export 'src/advanced_start_nw_tables.dart';
+export 'src/advanced_start_developable_resources.dart';
 export 'src/advanced_start_regiments.dart';
 export 'src/combat_config.dart';
 export 'src/regiment_type_display_name.dart';
