+++ packages/colonizethis_data/lib/src/unit_roles.dart
@@ -1,26 +1,21 @@
+import 'package:colonizethis_models/colonizethis_models.dart';
+
 import 'combat_config.dart';
 
 /// High-level unit roles used by validation and combat.
-enum UnitRole {
-  military,
-  civilianWorker,
-  explorer,
-  merchant,
-  spy,
-  naval,
-}
+enum UnitRole { military, civilianWorker, explorer, merchant, spy, naval }
 
 /// Explicit role mapping for non-military units.
 ///
 /// Military regiments are derived from [regimentStatsById] instead of being
 /// listed here.
 const Map<String, UnitRole> _explicitUnitRoles = {
-  'Explorer': UnitRole.explorer,
-  'Builder': UnitRole.civilianWorker,
-  'Engineer': UnitRole.civilianWorker,
-  'Spy': UnitRole.spy,
-  'Merchant': UnitRole.merchant,
-  'Rail Builder': UnitRole.civilianWorker,
+  kUnitTypeExplorer: UnitRole.explorer,
+  kUnitTypeBuilder: UnitRole.civilianWorker,
+  kUnitTypeEngineer: UnitRole.civilianWorker,
+  kUnitTypeSpy: UnitRole.spy,
+  kUnitTypeMerchant: UnitRole.merchant,
+  kUnitTypeRailBuilder: UnitRole.civilianWorker,
 };
 
 /// Returns the [UnitRole] for a given unit [type] id, or null if unknown.
@@ -35,20 +30,16 @@ UnitRole? unitRoleForType(String type) {
   return null;
 }
 
-bool isMilitaryUnit(String type) =>
-    unitRoleForType(type) == UnitRole.military;
+bool isMilitaryUnit(String type) => unitRoleForType(type) == UnitRole.military;
 
-bool isExplorerUnit(String type) =>
-    unitRoleForType(type) == UnitRole.explorer;
+bool isExplorerUnit(String type) => unitRoleForType(type) == UnitRole.explorer;
 
 bool isCivilianWorkerUnit(String type) =>
     unitRoleForType(type) == UnitRole.civilianWorker;
 
-bool isMerchantUnit(String type) =>
-    unitRoleForType(type) == UnitRole.merchant;
+bool isMerchantUnit(String type) => unitRoleForType(type) == UnitRole.merchant;
 
-bool isSpyUnit(String type) =>
-    unitRoleForType(type) == UnitRole.spy;
+bool isSpyUnit(String type) => unitRoleForType(type) == UnitRole.spy;
 
 bool isNavalUnit(String type) => unitRoleForType(type) == UnitRole.naval;
 
@@ -56,4 +47,3 @@ bool isNavalUnit(String type) => unitRoleForType(type) == UnitRole.naval;
 ///
 /// Currently only military units can initiate or participate in combat.
 bool canUnitInitiateCombat(String type) => isMilitaryUnit(type);
-
